multi-agent-be-feat-pdf-to-db/requirements.txt
alembic
annotated-types
anyio
bcrypt
blinker
certifi
chardet
click
colorama
distro
exceptiongroup
Flask
Flask-Cors
Flask-JWT-Extended
Flask-Migrate
Flask-SocketIO
Flask-SQLAlchemy
greenlet
groq
h11
httpcore
httpx
idna
itsdangerous
Jinja2
Mako
MarkupSafe
pdfminer3
psycopg2-binary
pycryptodome
pydantic
pydantic_core
PyJWT
python-dateutil
python-docx
python-dotenv
pytz
six
sniffio
sortedcontainers
SQLAlchemy
typing_extensions
tzdata
tzlocal
Werkzeug
uvicorn
langchain_groq
pandas